Crate problem116

Problem 116 (Red, green or blue tiles)

Solution detail

This problem is suited very well to a solution based on a linear recurrence. Let f(n, m) be the number of ways of tiling a row of length n with tiles of length m. We have the following choices:

Don't put anything in the first position, with f(n - 1, m) choices for the rest of the row.

Put a tile in the first position, with f(n - m, m) choices for the rest of the row.

This leads to the recurrence f(n, m) = f(n - 1, m) + f(n - m, m) with initial values given by f(0, m) = f(1, m) = ... = f(m - 1, m) = 1.

We seek the value of f(50, 2) + f(50, 3) + f(50, 4) - 3 - subtracting three because leaving and empty row is not allowed in each case.

Structs

BricksRecurrenceIterator
A structure capable of iterating over the number of ways of tiling a row of length n with tiles of length exactly m.

Functions

answer
Solve the problem, returning the answer as a String

main

solve
Find the number of ways of tiling a row of length n with red, green or blue bricks.